@@ -1,17 +1,16 @@
 apply plugin: 'com.android.library'
 
 android {
-    compileSdkVersion 27
-
+    compileSdkVersion 29
 
 
     defaultConfig {
         minSdkVersion 14
         targetSdkVersion 26
-        versionCode 10
-        versionName "2.0.0"
+        versionCode 21
+        versionName "2.1.0"
 
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
 //        externalNativeBuild {
 //            cmake {
@@ -41,13 +40,15 @@ android {
 }
 
 dependencies {
-    api 'com.android.support:appcompat-v7:27.1.1'
-    api 'com.android.support:support-v4:27.1.1'
+    api 'androidx.appcompat:appcompat:1.0.0'
+    api 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation fileTree(include: ['*.jar'], dir: 'libs')
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
     api 'com.squareup.retrofit2:retrofit:2.0.2'
     api 'com.squareup.retrofit2:converter-gson:2.0.2'
     implementation project(':toolkit')
-    compile 'cn.yipianfengye.android:zxing-library:2.2'
+    implementation 'cn.yipianfengye.android:zxing-library:2.2'
+    api 'com.tencent.tbs.tbssdk:sdk:43939'
+    implementation project(path: ':smartalert')
 }
