@@ -5,21 +5,19 @@ import android.content.DialogInterface;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.PopupMenu;
+import androidx.appcompat.app.AlertDialog;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.PopupMenu;
 import android.text.TextUtils;
 import android.view.MenuItem;
 import android.view.View;
 import android.webkit.JavascriptInterface;
-import android.webkit.WebView;
-import android.widget.Button;
-import android.widget.EditText;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import com.tencent.smtt.sdk.WebView;
 import com.zhuangfei.adapterlib.R;
 import com.zhuangfei.adapterlib.AdapterLibManager;
 import com.zhuangfei.adapterlib.utils.PackageUtils;
@@ -28,6 +26,10 @@ import com.zhuangfei.adapterlib.apis.TimetableRequest;
 import com.zhuangfei.adapterlib.apis.model.BaseResult;
 import com.zhuangfei.adapterlib.core.IArea;
 import com.zhuangfei.adapterlib.core.JsSupport;
+
+import java.util.HashMap;
+import java.util.Map;
+
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
@@ -60,11 +62,14 @@ public class UploadHtmlActivity extends AppCompatActivity {
     TextView displayTextView;
     public int nowIndex=0;
 
+    Map<String, String> extraHeaders = new HashMap<String, String>();
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_upload_html);
         ViewUtils.setStatusTextGrayColor(this);
+        extraHeaders.put("X-Requested-With", "");
         initView();
         initUrl();
         loadWebView();
@@ -112,24 +117,6 @@ public class UploadHtmlActivity extends AppCompatActivity {
      * 显示弹出菜单
      */
     public void showPopmenu() {
-        PopupMenu popup = new PopupMenu(this, helpView);
-        popup.getMenuInflater().inflate(R.menu.menu_webview, popup.getMenu());
-        popup.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
-            public boolean onMenuItemClick(MenuItem item) {
-                if(item.getItemId()==R.id.top1){
-                    String now=webView.getUrl();
-                    if(now.indexOf("/")!=-1){
-                        int index=now.lastIndexOf("/");
-                        webView.loadUrl(now.substring(0,index)+"/xkAction.do?actionType=6");
-                    }else{
-                        webView.loadUrl(now+"/xkAction.do?actionType=6");
-                    }
-                }
-                return true;
-            }
-        });
-
-        popup.show();
     }
 
 
@@ -143,6 +130,7 @@ public class UploadHtmlActivity extends AppCompatActivity {
         ua=ua.replace("Android","ndroidA");
 
         webView.getSettings().setUserAgentString(ua);
+//        webView.getSettings().setUserAgentString("Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/75.0.3770.80 Safari/537.36");
         webView.loadUrl(url);
     }
 
