@@ -1,13 +1,10 @@
 package com.zhuangfei.adapterlib.activity.scan;
 
-import android.Manifest;
 import android.content.Intent;
-import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
 import android.os.Bundle;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 import android.text.TextUtils;
 import android.view.View;
 import android.widget.LinearLayout;
@@ -17,6 +14,7 @@ import com.uuzuche.lib_zxing.activity.CaptureFragment;
 import com.uuzuche.lib_zxing.activity.CodeUtils;
 import com.zhuangfei.adapterlib.R;
 import com.zhuangfei.adapterlib.activity.AdapterSchoolActivity;
+import com.zhuangfei.adapterlib.utils.ViewUtils;
 
 public class ScanImportActivity extends AppCompatActivity {
 
