@@ -2,7 +2,7 @@ package com.zhuangfei.adapterlib.core;
 
 import android.annotation.SuppressLint;
 import android.app.Activity;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.text.TextUtils;
 import android.webkit.JavascriptInterface;
 
