@@ -1,7 +1,7 @@
 package com.zhuangfei.importcourse;
 
 import android.app.Activity;
-import android.content.Context;
+import android.app.AlertDialog;
 import android.content.Intent;
 import android.os.Bundle;
 import android.text.TextUtils;
@@ -9,18 +9,15 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.Toast;
 
-import com.tencent.stat.StatMultiAccount;
-import com.tencent.stat.StatService;
+import com.google.gson.Gson;
 import com.zhuangfei.adapterlib.AdapterLibManager;
-import com.zhuangfei.adapterlib.IStatSendCallback;
 import com.zhuangfei.adapterlib.ParseManager;
-import com.zhuangfei.adapterlib.StatManager;
-import com.zhuangfei.adapterlib.activity.SearchSchoolActivity;
+import com.zhuangfei.adapterlib.activity.AutoImportActivity;
+import com.zhuangfei.adapterlib.activity.NewSearchSchoolActivity;
 import com.zhuangfei.adapterlib.core.ParseResult;
+import com.zhuangfei.adapterlib.utils.GsonUtils;
 
 import java.util.List;
-import java.util.Map;
-import java.util.Properties;
 
 public class MainActivity extends Activity {
 
@@ -41,9 +38,9 @@ public class MainActivity extends Activity {
         btn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent=new Intent(MainActivity.this, SearchSchoolActivity.class);
-                intent.putExtra(SearchSchoolActivity.EXTRA_SEARCH_KEY,"河南理工大学");
+                Intent intent=new Intent(MainActivity.this, AutoImportActivity.class);
                 startActivityForResult(intent,1);
+                App.enter = true;
             }
         });
     }
@@ -51,14 +48,13 @@ public class MainActivity extends Activity {
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
-        if(requestCode==1&&resultCode==SearchSchoolActivity.RESULT_CODE){
+        if(requestCode==1&&resultCode== AutoImportActivity.RESULT_CODE){
             if(ParseManager.isSuccess()&&ParseManager.getData()!=null){
                 List<ParseResult> result=ParseManager.getData();
-                String str="";
-                for(ParseResult item:result){
-                    str+=item.getName()+"\n";
-                }
-                Toast.makeText(MainActivity.this,str, Toast.LENGTH_SHORT).show();
+                AlertDialog.Builder builder=new AlertDialog.Builder(this)
+                        .setTitle("Timetable")
+                        .setMessage(GsonUtils.getGson().toJson(result));
+                builder.create().show();
             }
         }
     }
