@@ -3,9 +3,14 @@ package com.zhuangfei.adapterlib;
 import android.content.Context;
 import android.text.TextUtils;
 import android.util.Log;
+
+import com.tencent.smtt.export.external.TbsCoreSettings;
+import com.tencent.smtt.sdk.QbSdk;
 import com.zhuangfei.adapterlib.apis.model.ValuePair;
 import com.zhuangfei.adapterlib.callback.OnValueCallback;
 import com.zhuangfei.adapterlib.callback.OnVersionFindCallback;
+
+import java.util.HashMap;
 
 /**
  * Created by Liu ZhuangFei on 2018/10/19.
@@ -14,8 +19,8 @@ public class AdapterLibManager {
     private static final String TAG = "AdapterLibManager";
 
     //核心库版本号
-    private static int libVersionNumber=10;
-    private static String libVersionName="lib-2.0.0";
+    private static int libVersionNumber=21;
+    private static String libVersionName="lib-2.1.0";
 
     public static int appVersionNumber=1;
 
@@ -42,6 +47,12 @@ public class AdapterLibManager {
     public static void register(String appkey, String stat){
         AdapterLibManager.appKey=appkey;
         AdapterLibManager.stat=stat;
+
+        // 在调用TBS初始化、创建WebView之前进行如下配置
+        HashMap map = new HashMap();
+        map.put(TbsCoreSettings.TBS_SETTINGS_USE_SPEEDY_CLASSLOADER, true);
+        map.put(TbsCoreSettings.TBS_SETTINGS_USE_DEXLOADER_SERVICE, true);
+        QbSdk.initTbsSettings(map);
     }
 
     public static void checkUpdate(final Context context,String updateId, final OnVersionFindCallback callback){
